+++ modules/Material/OverlayLayer.qml
@@ -46,9 +46,15 @@ Rectangle {
 
     MouseArea {
         anchors.fill: parent
-        enabled: overlayLayer.currentOverlay != null && overlayLayer.currentOverlay.globalMouseAreaEnabled
+        enabled: overlayLayer.currentOverlay != null && 
+                overlayLayer.currentOverlay.globalMouseAreaEnabled
         hoverEnabled: enabled
-        onClicked: if (overlayLayer.currentOverlay.dismissOnTap) overlayLayer.currentOverlay.close()
+        
         onWheel: wheel.accepted = true
+
+        onClicked: {
+            if (overlayLayer.currentOverlay.dismissOnTap)
+                overlayLayer.currentOverlay.close()
+        }        
     }
 }